+++ include/net/xfrm.h
@@ -148,7 +148,7 @@ struct xfrm_dev_offload {
 	struct net_device	*dev;
 	netdevice_tracker	dev_tracker;
 	struct net_device	*real_dev;
-	unsigned long		offload_handle;
+	uintptr_t		offload_handle;
 	u8			dir : 2;
 	u8			type : 2;
 	u8			flags : 2;
